requirements.txt
pytorch-lightning
PyYAML
tensorboard
torch
torchsummary
torchvision